+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+**Change Log v2.5.1**  
+- Add $SP.people() to find the user details based on a name
+- Add $SP.addressbook() to find someone in the Active Directory based on a part of his name
+- Check compatibility with jQuery 1.7.2
+
 **Change Log v2.5**  
 - Add compatibility with Sharepoint 2010  
 - Function "progress" added to $SP.list().add()